Functional test: Landing page and Join page now merged
