+++ frontend/app/views/index.scala.html
@@ -216,6 +216,7 @@
                         "Patrons play a vital role to keep the Guardian fearless, progressive and free from interference",
                         "Learn more about Patrons",
                         routes.Info.patron.toString,
+                        "qa-learn-patron",
                         image,
                         Some("tone-patron")
                     )
