+++ functional-tests/src/main/scala/com/gu/membership/pages/JoinFlowChooseTierPage.scala
@@ -2,9 +2,4 @@ package com.gu.membership.pages
 
 import org.openqa.selenium.WebDriver
 
-/**
- * Created by jao on 13/08/2014.
- */
-class JoinFlowChooseTierPage(driver: WebDriver) extends JoinPage(driver) {
-
-}
+class JoinFlowChooseTierPage(driver: WebDriver) extends LandingPage(driver) {}
