+++ functional-tests/src/main/scala/com/gu/membership/pages/PatronPage.scala
@@ -1,0 +1,14 @@
+package com.gu.membership.pages
+
+import org.openqa.selenium.{By, WebDriver}
+
+class PatronPage(driver: WebDriver) {
+
+  private def joinButton = driver.findElement(By.cssSelector(".qa-package-patron"))
+
+  def clickJoinButton: PaymentPage = {
+    joinButton.click()
+    new PaymentPage(driver)
+  }
+
+}
